+++ packages/runtime/src/runtime/header.tsx
@@ -2,6 +2,7 @@ import type { EnsembleHeaderModel } from "@ensembleui/react-framework";
 import { isObject } from "lodash-es";
 // eslint-disable-next-line import/no-cycle
 import { Column } from "../widgets/Column";
+import { WidgetRegistry } from "../registry";
 
 interface EnsembleHeaderProps {
   header?: EnsembleHeaderModel;
@@ -12,17 +13,26 @@ export const EnsembleHeader: React.FC<EnsembleHeaderProps> = ({ header }) => {
     return null;
   }
 
-  const titleWidget = isObject(header.title)
-    ? header.title
-    : {
-        name: "Text",
-        properties: {
-          text: header.title,
-          styles: {
-            color: header.styles?.titleColor || "black",
-          },
+  let titleWidget;
+  if (isObject(header.title)) {
+    titleWidget = header.title;
+  } else if (WidgetRegistry.findOrNull(header.title) !== null) {
+    titleWidget = {
+      name: header.title,
+      properties: {},
+    };
+  } else {
+    titleWidget = {
+      name: "Text",
+      properties: {
+        text: header.title,
+        styles: {
+          color: header.styles?.titleColor || "black",
+          ...header.styles,
         },
-      };
+      },
+    };
+  }
 
   return (
     <Column
@@ -31,7 +41,7 @@ export const EnsembleHeader: React.FC<EnsembleHeaderProps> = ({ header }) => {
         width: "100%",
         display: "flex",
         justifyContent: "center",
-        alignItems: header.styles?.centerTitle ? "center" : "normal",
+        alignItems: header.styles?.centerTitle === false ? "normal" : "center",
         backgroundColor: header.styles?.backgroundColor || "white",
         height: header.styles?.titleBarHeight || 56,
       }}